Fix non-determistic bug where components were sometimes not defined
The problem can be traced back to the jsonld library that sometimes fails to expand some IRIs,
in our case the IRIs of components, when complex context structures are in place.
The issue seems to be resolved by creating a new jsonld instance every time a new JSON document is parsed,
this effectively disables the caching of contexts which is possibly (part of) the cause.

More specifically, jsonld seems to fail on the part where it has to join a local context, contexts from URIs and the active context.
This joining happens using a cache, in which the bug is probably caused, as bypassing this caching mechanism resolves the problem.