Require URLs instead of literals when importing
Previously, the range of owl:imports was assumed to be a literal string that referred to a file path.
This made imports useless on the Web.

With this new change, the range of owl:imports is the set of URIs.
Together with this, a `lsd:importPaths` entry can be added to the package.json file.
This entry can contain a hash of mappings from URL prefix to local path.
For example:
```
  "lsd:importPaths": {
    "http://example.org/mystuff/components/": "components/"
  },
```

This is a breaking change and is not backwards-compatible with the old form of imports.