Remove required and unique param flags in favor of parameter ranges
This allows more complex ranges to be defined, such as nested arrays, or
arrays of union types.

Because of this change, arrays now must always be explicitly defined
within an RDF list (or @list in JSON-LD).
If one param value is provided, it is considered a singular value.
If the param value contains an RDF list, it is considered an array.
If multiple param values are provided without RDF list, an error is
thrown.