Support generic type instantiation during component extension
The general idea is that the extends range of a component can now also
be a component wrapped in an anonymous node together with certain
generic type instantiations.

When type-checking parameters for such generic types, such wrapped
extends ranges need to be properly handled, and may even lead to the
bindining of certain generics.

For example, if a parameter accepts Interface<T>, and we have a class
Class that implements Interface<boolean>. If a user then passes Class as
a value to this parameter within the config file, then the framework
will derive the binding of T to boolean.